Move ISO data classes to artifact openid-data-classes
