+++ CHANGELOG.md
@@ -13,7 +13,10 @@ Release 5.8.0:
  - Code organization:
    - Remove code elements deprecated in `5.7.0`
    - Remove all remaining `serialize()` and `deserialize()` methods in data classes
-   - Move data classes for token status into artifact `openid-data-classes`
+   - Move data classes for token status into artifact `openid-data-classes`, keeping the namespace
+ - Refactoring of ISO data classes:
+   - Move data classes from `vck` to `openid-data-classes`
+   - List of classes moved: `MobileSecurityObject`, `Document`, `IssuerSigned`, `DeviceResponse`
  - Issuer:
    - Extract interface `StatusListIssuer` out of `Issuer` to separate credential issuing and status list management
    - Extract interface `IssueCredentialFun` to be used in `CredentialIssuer` for OID4VCI
