+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/iso/DeviceResponse.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.lib.iso
+package at.asitplus.iso
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
