+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/CredentialWrapper.kt
@@ -1,8 +1,8 @@
 package at.asitplus.wallet.lib.agent.validation
 
+import at.asitplus.iso.IssuerSigned
 import at.asitplus.wallet.lib.data.VerifiableCredentialJws
 import at.asitplus.wallet.lib.data.VerifiableCredentialSdJwt
-import at.asitplus.wallet.lib.iso.IssuerSigned
 import kotlin.jvm.JvmInline
 
 sealed interface CredentialWrapper {
