+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/TokenStatusValidator.kt
@@ -1,12 +1,12 @@
 package at.asitplus.wallet.lib.agent.validation
 
+import at.asitplus.iso.IssuerSigned
 import at.asitplus.wallet.lib.agent.SubjectCredentialStore
 import at.asitplus.wallet.lib.data.Status
 import at.asitplus.wallet.lib.data.VerifiableCredentialJws
 import at.asitplus.wallet.lib.data.VerifiableCredentialSdJwt
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatusValidationResult
-import at.asitplus.wallet.lib.iso.IssuerSigned
 
 fun interface TokenStatusValidator {
     suspend operator fun invoke(status: Status): TokenStatusValidationResult
