+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/mdoc/MobileSecurityObjectTimelinessValidator.kt
@@ -1,9 +1,9 @@
 package at.asitplus.wallet.lib.agent.validation.mdoc
 
+import at.asitplus.iso.MobileSecurityObject
 import at.asitplus.wallet.lib.agent.validation.TimeScope
 import at.asitplus.wallet.lib.agent.validation.common.EntityExpiredError
 import at.asitplus.wallet.lib.agent.validation.common.EntityNotYetValidError
-import at.asitplus.wallet.lib.iso.MobileSecurityObject
 import io.github.aakira.napier.Napier
 
 class MobileSecurityObjectTimelinessValidator {
