+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/Deprecations.kt
@@ -1,0 +1,13 @@
+package at.asitplus.wallet.lib.iso
+
+@Deprecated("Renamed", ReplaceWith("at.asitplus.iso.MobileSecurityObject"))
+typealias MobileSecurityObject = at.asitplus.iso.MobileSecurityObject
+
+@Deprecated("Renamed", ReplaceWith("at.asitplus.iso.Document"))
+typealias Document = at.asitplus.iso.Document
+
+@Deprecated("Renamed", ReplaceWith("at.asitplus.iso.IssuerSigned"))
+typealias IssuerSigned = at.asitplus.iso.IssuerSigned
+
+@Deprecated("Renamed", ReplaceWith("at.asitplus.iso.DeviceResponse"))
+typealias DeviceResponse = at.asitplus.iso.DeviceResponse
