+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/iso/IsoProcessTest.kt
@@ -4,12 +4,16 @@ import at.asitplus.iso.DeviceAuth
 import at.asitplus.iso.DeviceKeyInfo
 import at.asitplus.iso.DeviceNameSpaces
 import at.asitplus.iso.DeviceRequest
+import at.asitplus.iso.DeviceResponse
 import at.asitplus.iso.DeviceSigned
 import at.asitplus.iso.DocRequest
+import at.asitplus.iso.Document
+import at.asitplus.iso.IssuerSigned
 import at.asitplus.iso.IssuerSignedItem
 import at.asitplus.iso.IssuerSignedList
 import at.asitplus.iso.ItemsRequest
 import at.asitplus.iso.ItemsRequestList
+import at.asitplus.iso.MobileSecurityObject
 import at.asitplus.iso.SingleItemsRequest
 import at.asitplus.iso.ValidityInfo
 import at.asitplus.iso.ValueDigest
